Refactor of inbox
