--- apps/frontend/src/app/features/personal/pages/inbox-page.component.scss
@@ -1,239 +1,0 @@
-:host {
-  display: block;
-}
-
-.page {
-  padding: 0.85rem 0 2rem;
-}
-
-.page-header h1 {
-  margin: 0;
-  font-size: clamp(2.7rem, 4vw, 3.6rem);
-  line-height: 1.02;
-  letter-spacing: -0.05em;
-}
-
-.page-header p {
-  margin: 0.55rem 0 0;
-  color: var(--on-surface-muted);
-  font-size: 1.12rem;
-}
-
-.capture-bar {
-  margin-top: 2rem;
-  display: grid;
-  grid-template-columns: auto minmax(0, 1fr) auto auto;
-  align-items: center;
-  gap: 0.8rem;
-  background: var(--surface-card);
-  border-radius: 1.2rem;
-  box-shadow:
-    inset 0 0 0 1px var(--outline-variant),
-    0 18px 32px var(--surface-dim);
-  padding: 0.8rem 0.95rem;
-}
-
-.capture-plus {
-  width: 2rem;
-  height: 2rem;
-  border: 0;
-  border-radius: 999px;
-  background: var(--primary-action-gradient);
-  color: hsl(var(--primary-foreground));
-  font-size: 1.4rem;
-  line-height: 1;
-}
-
-.capture-bar input {
-  border: 0;
-  outline: none;
-  background: transparent;
-  color: var(--on-surface);
-  font-size: 1.06rem;
-}
-
-.capture-bar input::placeholder {
-  color: var(--on-surface-subtle);
-}
-
-.capture-actions {
-  display: flex;
-  align-items: center;
-  gap: 0.6rem;
-  color: var(--on-surface-subtle);
-}
-
-.capture-project select {
-  border: 0;
-  border-radius: 999px;
-  background: var(--surface-container-highest);
-  color: var(--on-surface);
-  font-size: 0.92rem;
-  padding: 0.55rem 0.8rem;
-  min-width: 9rem;
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-}
-
-.capture-project select:focus {
-  outline: none;
-  box-shadow:
-    inset 0 0 0 1px var(--primary-solid),
-    0 0 0 3px var(--primary-soft);
-}
-
-.capture-submit {
-  border: 0;
-  border-radius: 0.92rem;
-  background: var(--primary-action-gradient);
-  color: hsl(var(--primary-foreground));
-  font-weight: 700;
-  min-width: 6.8rem;
-  min-height: 2.8rem;
-  padding: 0 1rem;
-}
-
-.capture-error,
-.status-copy {
-  margin: 0.7rem 0 0;
-  color: var(--on-surface-muted);
-}
-
-.capture-error {
-  color: #bb6d57;
-}
-
-.list-section {
-  margin-top: 2.8rem;
-}
-
-.task-stack {
-  display: grid;
-  gap: 0.95rem;
-}
-
-.empty-state {
-  border-radius: 1.4rem;
-  background: var(--surface-card);
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-  padding: 2rem 1.3rem;
-  text-align: center;
-}
-
-.empty-state h3 {
-  margin: 0;
-  font-size: 2rem;
-  letter-spacing: -0.04em;
-}
-
-.empty-state p {
-  margin: 0.55rem auto 0;
-  max-width: 28rem;
-  color: var(--on-surface-muted);
-}
-
-.promo-grid {
-  margin-top: 2rem;
-  display: grid;
-  grid-template-columns: minmax(0, 1.15fr) minmax(0, 0.85fr);
-  gap: 1rem;
-}
-
-.promo-card {
-  border-radius: 1.6rem;
-  padding: 1.5rem;
-}
-
-.promo-card-soft {
-  background: var(--surface-card);
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-}
-
-.promo-icon {
-  width: 4rem;
-  height: 4rem;
-  border-radius: 999px;
-  background: var(--primary-soft);
-  color: var(--primary-solid);
-  display: grid;
-  place-items: center;
-  font-size: 1.4rem;
-}
-
-.promo-card h3,
-.promo-card-dark p {
-  margin: 1rem 0 0;
-}
-
-.promo-card-soft p,
-.promo-card-dark span {
-  color: var(--on-surface-muted);
-}
-
-.promo-card-dark {
-  background: linear-gradient(
-    135deg,
-    var(--surface-container-highest),
-    var(--surface-container-high)
-  );
-  color: var(--on-surface);
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-}
-
-.promo-card-dark p {
-  font-size: 1.22rem;
-  line-height: 1.35;
-  font-style: italic;
-}
-
-.promo-card-dark span {
-  margin-top: 1rem;
-  color: var(--on-surface-subtle);
-  font-size: 0.82rem;
-  text-transform: uppercase;
-  letter-spacing: 0.18em;
-}
-
-@media (max-width: 900px) {
-  .promo-grid {
-    grid-template-columns: 1fr;
-  }
-}
-
-@media (max-width: 720px) {
-  .page {
-    padding-top: 0.3rem;
-  }
-
-  .capture-bar {
-    grid-template-columns: 1fr;
-    align-items: stretch;
-    gap: 0.55rem;
-    padding: 0.85rem;
-  }
-
-  .capture-plus {
-    width: 2.4rem;
-    height: 2.4rem;
-  }
-
-  .capture-actions {
-    display: grid;
-    grid-template-columns: auto minmax(0, 1fr);
-    align-items: center;
-    gap: 0.55rem;
-  }
-
-  .capture-project select {
-    width: 100%;
-  }
-
-  .capture-submit {
-    width: 100%;
-  }
-
-  .promo-grid {
-    grid-template-columns: 1fr;
-  }
-}
