--- apps/frontend/src/app/features/personal/pages/search-page.component.scss
@@ -1,222 +1,0 @@
-:host {
-  display: block;
-}
-
-.page {
-  padding: 1rem 0 2.5rem;
-}
-
-.header-chip {
-  border-radius: 999px;
-  background: var(--primary-soft);
-  color: var(--primary-solid);
-  padding: 0.55rem 0.8rem;
-  font-size: 0.84rem;
-  font-weight: 700;
-}
-
-.search-panel {
-  margin-top: 1.35rem;
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.search-input-shell {
-  flex: 1;
-  min-width: 0;
-  display: flex;
-  align-items: center;
-  gap: 0.55rem;
-  border-radius: 999px;
-  padding: 0.82rem 1rem;
-  background: var(--surface-overlay);
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-}
-
-.search-input-shell:focus-within {
-  box-shadow:
-    inset 0 0 0 1px rgba(35, 104, 73, 0.32),
-    0 10px 28px rgba(28, 28, 23, 0.05);
-}
-
-.search-input-shell input {
-  width: 100%;
-  border: 0;
-  outline: none;
-  background: transparent;
-  color: var(--on-surface);
-  font-size: 0.96rem;
-}
-
-.search-input-shell input::placeholder {
-  color: var(--on-surface-subtle);
-}
-
-.search-button {
-  border: 0;
-  border-radius: 999px;
-  background: var(--primary-action-gradient);
-  color: hsl(var(--primary-foreground));
-  font-size: 0.9rem;
-  font-weight: 700;
-  padding: 0.82rem 1.1rem;
-}
-
-.search-icon {
-  width: 1rem;
-  height: 1rem;
-  color: var(--on-surface-subtle);
-  flex: 0 0 auto;
-}
-
-.search-icon svg {
-  width: 100%;
-  height: 100%;
-  display: block;
-  stroke: currentColor;
-  fill: none;
-  stroke-width: 1.8;
-  stroke-linecap: round;
-  stroke-linejoin: round;
-}
-
-.tab-row {
-  margin-top: 1rem;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.6rem;
-}
-
-.tab-chip {
-  border-radius: 999px;
-  background: var(--surface-overlay);
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-  color: var(--on-surface-muted);
-  display: inline-flex;
-  align-items: center;
-  gap: 0.45rem;
-  padding: 0.68rem 0.95rem;
-  font-size: 0.84rem;
-  font-weight: 700;
-}
-
-.tab-chip strong {
-  border-radius: 999px;
-  padding: 0.15rem 0.45rem;
-  background: var(--primary-soft);
-  color: var(--primary-solid);
-}
-
-.tab-chip-active {
-  background: var(--primary-soft-strong);
-  color: var(--primary-solid);
-}
-
-.result-group {
-  margin-top: 1.35rem;
-}
-
-.task-stack {
-  display: grid;
-  gap: 0.9rem;
-}
-
-.project-grid {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 0.9rem;
-}
-
-.project-card {
-  display: grid;
-  gap: 0.8rem;
-  border-radius: 1.25rem;
-  background: var(--surface-card);
-  box-shadow: 0 14px 28px var(--surface-dim);
-  padding: 1.05rem 1.1rem;
-  color: inherit;
-  text-decoration: none;
-}
-
-.project-topline {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 1rem;
-}
-
-.project-topline h3 {
-  margin: 0;
-  font-size: 1.16rem;
-  letter-spacing: -0.03em;
-}
-
-.project-topline span {
-  color: var(--on-surface-subtle);
-  font-size: 0.82rem;
-  text-align: right;
-}
-
-.project-card p {
-  margin: 0;
-  color: var(--on-surface-muted);
-  line-height: 1.45;
-}
-
-.project-meta {
-  display: flex;
-  align-items: center;
-  gap: 0.55rem;
-  flex-wrap: wrap;
-  color: var(--on-surface-muted);
-  font-size: 0.8rem;
-  font-weight: 700;
-}
-
-.project-meta span {
-  border-radius: 999px;
-  background: var(--surface-container-low);
-  padding: 0.25rem 0.55rem;
-}
-
-.empty-state {
-  margin-top: 1.25rem;
-  border-radius: 1.4rem;
-  box-shadow: inset 0 0 0 1px var(--outline-variant);
-  background: var(--surface-card);
-  padding: 1.6rem 1.3rem;
-}
-
-.empty-state.compact {
-  margin-top: 0;
-}
-
-.empty-state h2 {
-  margin: 0;
-  font-size: 1.5rem;
-  letter-spacing: -0.04em;
-}
-
-.empty-state p {
-  margin: 0.55rem 0 0;
-  color: var(--on-surface-muted);
-  line-height: 1.45;
-}
-
-@media (max-width: 840px) {
-  .project-grid {
-    grid-template-columns: 1fr;
-  }
-}
-
-@media (max-width: 720px) {
-  .search-panel {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .search-button {
-    width: 100%;
-  }
-}
