+++ apps/frontend/src/app/features/personal/shell/personal-shell.component.html
@@ -16,9 +16,10 @@
       </div>
 
       <nav class="nav-list" aria-label="Personal navigation">
-        @for (item of navItems; track item.route) {
+        @for (item of navItems; track item.route + (item.queryParams?.['view'] || '')) {
           <a
             [routerLink]="item.route"
+            [queryParams]="item.queryParams"
             routerLinkActive="nav-item-active"
             [routerLinkActiveOptions]="{ exact: true }"
             class="nav-item"
@@ -52,7 +53,12 @@
     </div>
 
     <div class="sidebar-bottom">
-      <a routerLink="/inbox" fragment="capture" class="quick-add-button">
+      <a
+        [routerLink]="['/tasks']"
+        [queryParams]="{ view: 'inbox' }"
+        fragment="capture"
+        class="quick-add-button"
+      >
         <span class="quick-add-icon"><fa-icon [icon]="faPlus"></fa-icon></span>
         Quick Add Task
       </a>
@@ -77,7 +83,7 @@
           <fa-icon [icon]="faBars" aria-hidden="true"></fa-icon>
         </button>
 
-        <a routerLink="/inbox" class="topbar-brand">
+        <a [routerLink]="['/tasks']" [queryParams]="{ view: 'inbox' }" class="topbar-brand">
           <span class="topbar-brand-mark" aria-hidden="true">
             <svg viewBox="0 0 24 24">
               <path
